# perf(turbo-tasks): Use native async traits for TaskInput (not the async_trait macro)