--- a/turbopack/crates/turbo-tasks/src/task/task_input.rs
+++ b/turbopack/crates/turbo-tasks/src/task/task_input.rs
@@ -1,7 +1,6 @@
-use std::{any::Any, fmt::Debug, hash::Hash};
+use std::{any::Any, fmt::Debug, future::Future, hash::Hash};
 
 use anyhow::Result;
-use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 
 use crate::{
@@ -12,10 +11,9 @@ use crate::{
 /// [`#[turbo_tasks::function]`][crate::function] argument.
 ///
 /// See also [`ConcreteTaskInput`].
-#[async_trait]
 pub trait TaskInput: Send + Sync + Clone + Debug + PartialEq + Eq + Hash {
-    async fn resolve(&self) -> Result<Self> {
-        Ok(self.clone())
+    fn resolve(&self) -> impl Future<Output = Result<Self>> + Send + '_ {
+        async { Ok(self.clone()) }
     }
     fn is_resolved(&self) -> bool {
         true
@@ -28,7 +26,6 @@ pub trait TaskInput: Send + Sync + Clone + Debug + PartialEq + Eq + Hash {
 macro_rules! impl_task_input {
     ($($t:ty),*) => {
         $(
-            #[async_trait]
             impl TaskInput for $t {}
         )*
     };
@@ -48,7 +45,6 @@ impl_task_input! {
     ValueTypeId
 }
 
-#[async_trait]
 impl<T> TaskInput for Vec<T>
 where
     T: TaskInput,
@@ -70,7 +66,6 @@ where
     }
 }
 
-#[async_trait]
 impl<T> TaskInput for Option<T>
 where
     T: TaskInput,
@@ -97,7 +92,6 @@ where
     }
 }
 
-#[async_trait]
 impl<T> TaskInput for Vc<T>
 where
     T: Send,
@@ -220,7 +214,6 @@ impl<'de, T> Deserialize<'de> for TransientInstance<T> {
 
 macro_rules! tuple_impls {
     ( $( $name:ident )+ ) => {
-        #[async_trait]
         impl<$($name: TaskInput),+> TaskInput for ($($name,)+)
         where $($name: TaskInput),+
         {